Prefer wider types over narrowing casts in core/vnl
Instead of narrowing via static_cast, prefer keeping the source type
(size_t, unsigned int, long) and cascading type changes to downstream
variables (loop counters, local vars). Where narrowing casts remain
necessary (public API boundaries), add assert() guards to verify values
fit in the target type before casting.

Key changes:
- vnl_error.h: Add inline overloads accepting unsigned int, size_t to
  avoid casts at call sites; forward to original int-based functions
  after asserting values are in range
- vnl_amoeba.cxx: Use size_t n from x.size(), cascade to loop vars;
  assert before static_cast<int> at SimplexCorner constructor calls
- vnl_powell.cxx: Use size_t n, add size_t constructor overload for
  local vnl_powell_1dfun struct, cascade loop vars to size_t
- vnl_gaussian_kernel_1d: Use size_t for wid and loop var; add assert
  in width() before narrowing return
- vnl_lbfgs.cxx: Use long for l and loop vars to match long n (from
  Fortran interface); use constexpr for compile-time constants
- vnl_levenberg_marquardt.cxx: Change inner loop var to int to match
  vnl_vector<int> jpvt, eliminating the cast
- vnl_planefit.cxx: Use unsigned int for npts/dim to match matrix
  rows()/columns() return type
- vnl_cholesky.h, vnl_ldl_cholesky.h, vnl_nonlinear_minimizer.h:
  Add assert guards before remaining narrowing casts in getters
- Remove static_casts at vnl_error call sites in vnl_matrix.hxx,
  vnl_matrix_fixed.hxx, vnl_matrix_fixed_ref.hxx, vnl_vector.hxx,
  vnl_vector_fixed.h, vnl_vector_fixed.hxx, vnl_vector.h

+++ core/vnl/examples/vnl_planefit.cxx
@@ -18,12 +18,12 @@ main()
   std::cin >> pts;
 
   // Build design matrix D
-  const int npts = static_cast<int>(pts.rows());
-  const int dim = static_cast<int>(pts.columns());
+  const unsigned int npts = pts.rows();
+  const unsigned int dim = pts.columns();
   vnl_matrix<double> D(npts, dim + 1);
-  for (int i = 0; i < npts; ++i)
+  for (unsigned int i = 0; i < npts; ++i)
   {
-    for (int j = 0; j < dim; ++j)
+    for (unsigned int j = 0; j < dim; ++j)
       D(i, j) = pts(i, j);
     D(i, dim) = 1;
   }
